Address possible loss of data warnings (#107)
From Visual Studio, for example:

_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(160): warning C4018: '<': signed/unsigned mismatch
flag
Build Log Line	24
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(316): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
flag
Build Log Line	25
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(433): warning
C4244: 'return': conversion from 'double' to 'float', possible loss of
data

+++ src/lib/cleaver/Status.h
@@ -10,7 +10,7 @@
 
 class Status {
   public:
-    Status(int tot) : total_(double(tot)), count_(0), percent_(-1) {
+    Status(size_t tot) : total_(double(tot)), count_(0), percent_(-1) {
     } ;
     ~Status() {}
     void printStatus() {
